requirements.txt
requests
python-dotenv
mwcleric
discord
